+++ consensus/pacemaker.go
@@ -79,7 +79,7 @@ func (pm *pacemaker) newProposal() {
 
 	pm.driver.updateQCHigh(blk.QuorumCert())
 
-	var quota uint32 = 1
+	var quota uint64 = 1
 	if !TwoPhaseBFTFlag {
 		var err error
 		if quota, err = pm.status.getVoteQuota(); err != nil {
